This commit has a lot of changes which affect the base of cisTEM, including changes to the database and job launching (run profiles are now more complex, and allow threading. Only 1 job is now needed as the first job is now both a master and a slave).  This version will NOT OPEN databases from previous versions.
Big feature changes are true integration of refineCTF with beam tilt estimation, support for tilted data during CTF estimation and particle extraction. Full incorporation of template matching into the GUI.

This is very likely to be unstable and full of bugs (sorry).

Makefile.am - added combine_via_max, estimate_beamtilt and remove_relion_stripes.  Also moved refine_ctf out of experimental
cistem_parameters.cpp - fixed bug in ReplaceNanAndInfWithOther.
core_headers.h - added template_matching.h
database.h/cpp - Added ReturnIntArrayFromSelectCommand and added a number of things for template matching.  Added entries for tilt angle and axis to CTF estimation.
defines.h - Changed MyPrintWithDetails to print a stack trace when compiled in debug mode.
functions.h/cpp - Added functions for sending template matching results
gui_core_headers.h - Added files for Template matching / refineCTF
image.cpp - changed ReturnSumofSquares to use powf
job_packager.cpp - Added in details for sending the new syle run profile
job_tracker.h - ReturnPercentCompleted will now never return a number outside of 0-100
myapp.h/cpp - Many changes that allow the first job launched to act as both a master and a slave, this was necessary from the perspective of adding threading.
particle.cpp - Changed the check in InitCTF for minimum change required to regenerate the CTF from 40A to 1A.
progressbar.h/cpp - Added a virtual function CallOnUpdate, which is called during update. This allows people to add their own code into the progress bar.
project.h/cpp - Added Directories for template matching
reconstructed_volume.cpp - just added a print out of the size when a size error is thrown.
rle3d.cpp - removal of old commented out code
run_profiles.h/cpp - expanded run profile to allow for threading, also to overide the number of calculated jobs, this allows running a single command which itself launched many commands (e.g. using openmpi)
socket_codes.h - added template match result socket code
socket_communicator.h/cpp -  wrapped sending of the job package into YouAreTheMaster, this fixes a potential bug where the jobs could try and use the job package before they had it. Added some safety checks upon shutdown.
template_matching.h/cpp - classes to hold template matching results
AbInito3dPanel.cpp - Changes to run checks to allow run profiles with only 1 entry to run.  Added correct_ewald_sphere(hard coded to 0) to SetupReconstruction
AlignMovies.cpp - Changes to run checks to allow run profiles with only 1 entry to run. 
AutoRefine3dPanel.cpp - Added use of a signed_cc limit on odd rounds (but never the final round).  I hope this will improve overfitting, but we need to keep an eye open that this does not make results worse.  Changes to run checks to allow run profiles with only 1 entry to run.  Added correct_ewald_sphere(hard coded to 0) to SetupReconstruction
DisplayPanel.h/cpp - Added functions for showing template matching results, Also changed "ChangeFile" to allow changing of the non active tab, which should lead to reduced flickering (when implemented, it is only implemented in the TemplateMatching Results at the moment)
FindCTFPanel.cpp - Added flags / result handling for tilted searching, Changes to run checks to allow run profiles with only 1 entry to run.  
FindParticlesPanel.cpp - Changes to run checks to allow run profiles with only 1 entry to run. 
Generate3DPanel.h/cpp - Added flags for setting Ewald sphere correction. Changes to run checks to allow run profiles with only 1 entry to run.  
gui_functions.h/cpp - added delete refinectf_scratch
JobPanel.cpp - Changes to handle the fact that only 1 job is now needed.
MainFrame.h/cpp - Added support for all the new panels.
MatchTemplatePanel.h/cpp - Made it work properly with all options / results / database integration etc.
MatchTemplateResultsPanel.h/cpp - New panel for handling template matching results.
mathplot.h/cpp - Added ability to set minium step for X-axis plotting.
my_controls.cpp - expanded refinement results table to include per particle things
MyAddRunCommandDialog.h/cpp - changes for expanded run profiles
MyFindCTFResultsPanel.cpp - Added tilt angle / axis.
MyNewRefinementPackageWizard.cpp - Added support for tilted images, so that the defocus of each particle is calculated based on the tilt. Set the output pixel size precision to 4
MyRefine2DPanel.cpp - Set AutoCentre to yes as a default (still not automasking). Changes to run checks to allow run profiles with only 1 entry to run.
MyRefine3DPanel.cpp - Changes to run checks to allow run profiles with only 1 entry to run.  Added correct_ewald_sphere(hard coded to 0) to SetupReconstruction
MyRunProfilesPanel.cpp - Changes to handle the new style run profile.
PlotCurvePanel.h/cpp - Changes to allow a minimum plot step on the x-axis, and the ability to have y centred on the x;
RefineCTFPanel.h/cpp - Should now be useable, including beam tilt.
RefinementParametersDialog.cpp - Added per particle fields.
RefineTemplatePanel.h/cpp - New panel for refining the results of a match template
ShowCTFResultsPanel.h/cpp - Added tilt results
ShowTemplateMatchResultsPanel.h/cpp - Result Panel for Template matching
combine_via_max.cpp - Combine two files taking the max from each. useful for merging two overlapping maps.
ctffind.cpp - added result handling for tilts back to the GUI
estimate_beamtilt.cpp - new program for estimating beam tilt from a phase difference image. Only used by the GUI
guix_job_control.cpp - New socket methods, fixes for new 1 job minimum
match_template.cpp - Changes for GUI integration.
prepare_stack_class_average.cpp - use Best2DClass column instead of image is active
projectx.cpp - added all the new panels
reconstruct3d.cpp - exposed Ewald sphere correction as an input
refine_ctf.cpp - changes to allow threading to passed as a command line option
refine_template.cpp - Changes for GUI integration.
refine2d.cpp - changes to allow threading to passed as a command line option
refine3d.cpp - changes to allow threading to passed as a command line option
remove_inf_and_nan.cpp - now reports how many images had inf or nan
remove_relion_stripes.cpp - little tool to remove out the weird repeating pixel stripes relion puts in when cutting from outside the image.
unblur.cpp - changes to allow threading to passed as a command line option

+++ src/core/image.cpp
@@ -247,7 +247,7 @@ float Image::ReturnSumOfSquares(float wanted_mask_radius, float wanted_center_x,
 						{
 							if (distance_from_center_squared > mask_radius_squared)
 							{
-								sum += powf(real_values[address],2);
+								sum += pow(real_values[address],2);
 								number_of_pixels++;
 							}
 						}
@@ -255,7 +255,7 @@ float Image::ReturnSumOfSquares(float wanted_mask_radius, float wanted_center_x,
 						{
 							if (distance_from_center_squared <= mask_radius_squared)
 							{
-								sum += powf(real_values[address],2);
+								sum += pow(real_values[address],2);
 								number_of_pixels++;
 							}
 						}
@@ -281,7 +281,7 @@ float Image::ReturnSumOfSquares(float wanted_mask_radius, float wanted_center_x,
 				{
 					for (i = 0; i < logical_x_dimension; i++)
 					{
-						sum += powf(real_values[address],2);
+						sum += pow(real_values[address],2);
 						address++;
 					}
 					address += padding_jump_value;
